+++ packages/sdk/src/common/marketplace.ts
@@ -1,22 +1,22 @@
-import { BigNumber, BigNumberish, Contract, providers } from "ethers";
 import {
   InterfaceId_IERC1155,
   InterfaceId_IERC721,
 } from "../constants/contract";
-import { ContractWrapper } from "../core/classes/contract-wrapper";
 import { SignerOrProvider } from "../core";
+import { ContractWrapper } from "../core/classes/contract-wrapper";
+import { MAX_BPS } from "../schema/shared";
 import {
   NewAuctionListing,
   NewDirectListing,
   Offer,
 } from "../types/marketplace";
-import invariant from "tiny-invariant";
 import { fetchCurrencyValue } from "./currency";
-import { MAX_BPS } from "../schema/shared";
 import { IERC1155, IERC165, IERC721 } from "@thirdweb-dev/contracts-js";
-import ERC1155Abi from "@thirdweb-dev/contracts-js/abis/IERC1155.json";
-import ERC721Abi from "@thirdweb-dev/contracts-js/abis/IERC721.json";
-import ERC165Abi from "@thirdweb-dev/contracts-js/abis/IERC165.json";
+import ERC165Abi from "@thirdweb-dev/contracts-js/dist/abis/IERC165.json";
+import ERC721Abi from "@thirdweb-dev/contracts-js/dist/abis/IERC721.json";
+import ERC1155Abi from "@thirdweb-dev/contracts-js/dist/abis/IERC1155.json";
+import { BigNumber, BigNumberish, Contract, providers } from "ethers";
+import invariant from "tiny-invariant";
 
 /**
  * This method checks if the given token is approved for the transferrerContractAddress contract.